feat:optimize metadata module.
