+++ spring-cloud-starter-tencent-metadata-transfer/pom.xml
@@ -48,6 +48,7 @@
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-cloud-starter-openfeign</artifactId>
+			<optional>true</optional>
 		</dependency>
 
 		<!-- powermock-module-junit4 -->
