+++ spring-cloud-starter-tencent-metadata-transfer/src/main/java/com/tencent/cloud/metadata/core/interceptor/Metadata2HeaderRestTemplateInterceptor.java
@@ -15,70 +15,71 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package com.tencent.cloud.metadata.core.filter.gateway.zuul;
+package com.tencent.cloud.metadata.core.interceptor;
 
+import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
 import java.util.Map;
 
-import com.netflix.zuul.ZuulFilter;
-import com.netflix.zuul.context.RequestContext;
 import com.tencent.cloud.common.constant.MetadataConstant;
 import com.tencent.cloud.common.metadata.MetadataContext;
 import com.tencent.cloud.common.metadata.MetadataContextHolder;
 import com.tencent.cloud.common.util.JacksonUtils;
 
+import org.springframework.core.Ordered;
+import org.springframework.http.HttpRequest;
+import org.springframework.http.client.ClientHttpRequestExecution;
+import org.springframework.http.client.ClientHttpRequestInterceptor;
+import org.springframework.http.client.ClientHttpResponse;
 import org.springframework.util.CollectionUtils;
-
-import static org.springframework.cloud.netflix.zuul.filters.support.FilterConstants.RIBBON_ROUTING_FILTER_ORDER;
-import static org.springframework.cloud.netflix.zuul.filters.support.FilterConstants.ROUTE_TYPE;
+import org.springframework.util.StringUtils;
 
 /**
- * Zuul filter used for writing metadata in HTTP request header.
+ * Interceptor used for adding the metadata in http headers from context when web client
+ * is RestTemplate.
  *
  * @author Haotian Zhang
  */
-public class Metadata2HeaderZuulFilter extends ZuulFilter {
+public class Metadata2HeaderRestTemplateInterceptor
+		implements ClientHttpRequestInterceptor, Ordered {
 
 	@Override
-	public String filterType() {
-		return ROUTE_TYPE;
+	public int getOrder() {
+		return MetadataConstant.OrderConstant.METADATA_2_HEADER_INTERCEPTOR_ORDER;
 	}
 
 	@Override
-	public int filterOrder() {
-		return RIBBON_ROUTING_FILTER_ORDER - 1;
-	}
-
-	@Override
-	public boolean shouldFilter() {
-		return true;
-	}
-
-	@Override
-	public Object run() {
-		// get request context
-		RequestContext requestContext = RequestContext.getCurrentContext();
-
+	public ClientHttpResponse intercept(HttpRequest httpRequest, byte[] bytes,
+			ClientHttpRequestExecution clientHttpRequestExecution) throws IOException {
 		// get metadata of current thread
 		MetadataContext metadataContext = MetadataContextHolder.get();
 
 		// add new metadata and cover old
+		String metadataStr = httpRequest.getHeaders()
+				.getFirst(MetadataConstant.HeaderName.CUSTOM_METADATA);
+		if (!StringUtils.isEmpty(metadataStr)) {
+			Map<String, String> headerMetadataMap = JacksonUtils
+					.deserialize2Map(metadataStr);
+			for (String key : headerMetadataMap.keySet()) {
+				metadataContext.putTransitiveCustomMetadata(key,
+						headerMetadataMap.get(key));
+			}
+		}
 		Map<String, String> customMetadata = metadataContext
 				.getAllTransitiveCustomMetadata();
 		if (!CollectionUtils.isEmpty(customMetadata)) {
-			String metadataStr = JacksonUtils.serialize2Json(customMetadata);
+			metadataStr = JacksonUtils.serialize2Json(customMetadata);
 			try {
-				requestContext.addZuulRequestHeader(
-						MetadataConstant.HeaderName.CUSTOM_METADATA,
+				httpRequest.getHeaders().set(MetadataConstant.HeaderName.CUSTOM_METADATA,
 						URLEncoder.encode(metadataStr, "UTF-8"));
 			}
 			catch (UnsupportedEncodingException e) {
-				requestContext.addZuulRequestHeader(
-						MetadataConstant.HeaderName.CUSTOM_METADATA, metadataStr);
+				httpRequest.getHeaders().set(MetadataConstant.HeaderName.CUSTOM_METADATA,
+						metadataStr);
 			}
 		}
-		return null;
+		return clientHttpRequestExecution.execute(httpRequest, bytes);
 	}
 
 }
