+++ spring-cloud-tencent-commons/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,3 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
   com.tencent.cloud.common.util.ApplicationContextAwareUtils,\
-  com.tencent.cloud.common.metadata.MetadataAutoConfiguration
+  com.tencent.cloud.common.metadata.config.MetadataAutoConfiguration
