+++ spring-cloud-tencent-commons/src/test/java/com/tencent/cloud/common/metadata/config/MetadataLocalPropertiesTest.java
@@ -15,70 +15,49 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package com.tencent.cloud.metadata.core.filter.web;
+package com.tencent.cloud.common.metadata.config;
 
-import com.tencent.cloud.common.constant.MetadataConstant;
-import com.tencent.cloud.common.metadata.MetadataLocalProperties;
 import org.assertj.core.api.Assertions;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import reactor.core.publisher.Mono;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.mock.http.server.reactive.MockServerHttpRequest;
-import org.springframework.mock.web.server.MockServerWebExchange;
 import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.web.server.ServerWebExchange;
-import org.springframework.web.server.WebFilterChain;
-
-import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.MOCK;
 
 /**
- * Test for {@link MetadataReactiveFilter}
+ * Test for {@link MetadataLocalProperties}
  *
  * @author Haotian Zhang
  */
 @RunWith(SpringRunner.class)
-@SpringBootTest(webEnvironment = MOCK,
-		classes = MetadataServletFilterTest.TestApplication.class,
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
+		classes = MetadataLocalPropertiesTest.TestApplication.class,
 		properties = { "spring.config.location = classpath:application-test.yml" })
-public class MetadataReactiveFilterTest {
+public class MetadataLocalPropertiesTest {
 
 	@Autowired
 	private MetadataLocalProperties metadataLocalProperties;
 
-	private MetadataReactiveFilter metadataReactiveFilter;
-
-	@Before
-	public void setUp() {
-		this.metadataReactiveFilter = new MetadataReactiveFilter();
-	}
-
 	@Test
 	public void test1() {
-		Assertions.assertThat(this.metadataReactiveFilter.getOrder())
-				.isEqualTo(MetadataConstant.OrderConstant.FILTER_ORDER);
-	}
-
-	@Test
-	public void test2() {
-		// Create mock WebFilterChain
-		WebFilterChain webFilterChain = serverWebExchange -> Mono.empty();
-
-		// Mock request
-		MockServerHttpRequest request = MockServerHttpRequest.get("test")
-				.header(MetadataConstant.HeaderName.CUSTOM_METADATA, "{\"c\": \"3\"}")
-				.build();
-		ServerWebExchange exchange = MockServerWebExchange.from(request);
-
-		metadataReactiveFilter.filter(exchange, webFilterChain);
 		Assertions.assertThat(metadataLocalProperties.getContent().get("a"))
 				.isEqualTo("1");
 		Assertions.assertThat(metadataLocalProperties.getContent().get("b"))
 				.isEqualTo("2");
 		Assertions.assertThat(metadataLocalProperties.getContent().get("c")).isNull();
+	}
+
+	@Test
+	public void test2() {
+		Assertions.assertThat(metadataLocalProperties.getTransitive().contains("b"))
+				.isTrue();
+	}
+
+	@SpringBootApplication
+	protected static class TestApplication {
+
 	}
 
 }
