+++ spring-cloud-tencent-commons/src/test/resources/application-test.yml
@@ -1,6 +1,10 @@
 spring:
   application:
     name: test
+  autoconfigure:
+    exclude:
+      - org.springframework.cloud.gateway.config.GatewayAutoConfiguration
+      - org.springframework.cloud.netflix.zuul.ZuulServerAutoConfiguration
   cloud:
     tencent:
       metadata:
