+++ spring-cloud-tencent-examples/polaris-circuitbreaker-example/polaris-circuitbreaker-example-b/pom.xml
@@ -11,12 +11,9 @@
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>polaris-circuitbreaker-example-b</artifactId>
+	<name>Polaris Circuit Breaker Example B</name>
 
-    <properties>
-        <java.version>1.8</java.version>
-    </properties>
-
-    <dependencies>
+	<dependencies>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-webflux</artifactId>
@@ -51,5 +48,4 @@
             </plugin>
         </plugins>
     </build>
-
 </project>
