+++ spring-cloud-tencent-examples/polaris-discovery-example/discovery-callee-service/src/main/java/com/tencent/cloud/polaris/discovery/service/callee/DiscoveryCalleeController.java
@@ -17,6 +17,10 @@
 
 package com.tencent.cloud.polaris.discovery.service.callee;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -31,13 +35,19 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/discovery/service/callee")
 public class DiscoveryCalleeController {
 
+	private static Logger LOG = LoggerFactory.getLogger(DiscoveryCalleeController.class);
+
+	@Value("${server.port:0}")
+	private int port;
+
 	/**
 	 * Get information of callee.
 	 * @return information of callee
 	 */
 	@GetMapping("/info")
 	public String info() {
-		return "Discovery Service Callee";
+		LOG.info("Discovery Service Callee [{}] is called.", port);
+		return String.format("Discovery Service Callee [%s] is called.", port);
 	}
 
 	/**
@@ -48,6 +58,7 @@ public class DiscoveryCalleeController {
 	 */
 	@GetMapping("/sum")
 	public int sum(@RequestParam int value1, @RequestParam int value2) {
+		LOG.info("Discovery Service Callee is called and sum is {}.", value1 + value2);
 		return value1 + value2;
 	}
