+++ spring-cloud-tencent-examples/polaris-discovery-example/discovery-callee-service/src/main/resources/polaris.yml
@@ -1,3 +1,9 @@
+global:
+  system:
+    discoverCluster:
+      sameAsBuiltin: true
+    healthCheckCluster:
+      sameAsBuiltin: true
 #global:
 #  system:
 #    discoverCluster:
