+++ spring-cloud-tencent-examples/polaris-discovery-example/discovery-caller-service/pom.xml
@@ -23,21 +23,21 @@
 			<artifactId>spring-cloud-starter-tencent-polaris-discovery</artifactId>
 		</dependency>
 
-        <!--        <dependency>-->
-        <!--            <groupId>org.springframework.cloud</groupId>-->
-        <!--            <artifactId>spring-cloud-starter-consul-discovery</artifactId>-->
-        <!--        </dependency>-->
+		<!--        <dependency>-->
+		<!--            <groupId>org.springframework.cloud</groupId>-->
+		<!--            <artifactId>spring-cloud-starter-consul-discovery</artifactId>-->
+		<!--        </dependency>-->
 
-        <!--        <dependency>-->
-        <!--            <groupId>org.springframework.cloud</groupId>-->
-        <!--            <artifactId>spring-cloud-starter-netflix-eureka-client</artifactId>-->
-        <!--        </dependency>-->
+		<!--        <dependency>-->
+		<!--            <groupId>org.springframework.cloud</groupId>-->
+		<!--            <artifactId>spring-cloud-starter-netflix-eureka-client</artifactId>-->
+		<!--        </dependency>-->
 
-        <!--        <dependency>-->
-        <!--            <groupId>com.tencent.cloud</groupId>-->
-        <!--            <artifactId>spring-cloud-starter-tencent-polaris-router</artifactId>-->
-        <!--        </dependency>-->
-    </dependencies>
+		<dependency>
+			<groupId>com.tencent.cloud</groupId>
+			<artifactId>spring-cloud-starter-tencent-polaris-router</artifactId>
+		</dependency>
+	</dependencies>
     <build>
         <plugins>
             <plugin>
